Derive default manually
Implementing default manually avoids the hidden requirement that P: Default,
which is not implemented for some std types.

The use case is to allow the priority queue to be easily added to
already-default structures that derive default themselves as:

        #[derive(Default)]
        struct MyState {
            ...
            pub pq: PriorityQueue<i32, Instant>,
        }

Without needing to `impl Default` for the entire struct.

In my specific case, I use the priority queue as a callback mechanism
using time instants from std::time::Instant. If the top elements are
behind the `now()` time, we run these instants and put them back onto
the queue with an added interval (for repeating timers).