test: Changing priority does not change the contents
When we change a priority, I do not want to change the contents of the
element (do not basically want to pop and then push the new thing),
because sometimes we have elements (like callbacks in this test) that we
want to keep.

This patch implements a test that confirms this.